Implement a quick and dirty detail view logic This currently fills the detail view with the current weather by a for each loop through a list of locationmodels retrieved from a file. This seems ugly, and I should just save each locationmodel as a separate file by iD. It also does not save the forecast, and does not get any information on whether a refresh should be called.

+++ app/src/main/java/gro/gibberish/drizzle/models/LocationForecastModel.java
@@ -67,10 +67,10 @@ public class LocationForecastModel implements BaseModel, Serializable{
     }
 
     public static class Temp implements Serializable {
-        private int max;
-        private int min;
+        private double max;
+        private double min;
 
-        public int getMax() {
+        public double getMax() {
             return max;
         }
 
@@ -78,7 +78,7 @@ public class LocationForecastModel implements BaseModel, Serializable{
             this.max = max;
         }
 
-        public int getMin() {
+        public double getMin() {
             return min;
         }
